Preserve size information in diff mode
To perform a diff, Bloaty would previously subtract the sizes in the
base file's Rollup from the current file's and pass the resulting
Rollup to be unrolled and printed. This had the unfortunate side effect
of discarding the new and old size data for each label, leaving only the
size difference. This size data is useful when labels in a binary have
capacities associated with them.

This change modifies the way Bloaty performs diffs by moving the size
subtraction further along in the process, when creating RollupRows.
Original and current sizes are added to the RollupRow struct and set
when each row is created to preserve all size information for each
label. Bloaty's default pretty-printed output is unchanged, but the
[CT]SV output is modified to include additional columns for original
and current label sizes.