Fix deleted entries in diff mode
Commit 22efbd0 changed the way subtraction was performed in diff mode by
no longer using Rollup::Subtract. This had the side-effect of preventing
entries for deleted labels from being created in the new Rollup, causing
the output not to include any deleted labels.

This commit fixes the issue by adding in a function that creates the
missing entries in the same way that Rollup::Subtract did.